build(app): fix for new naming convention
