+++ src/app/app-routing-module.ts
@@ -1,4 +1,5 @@
-import { Routes } from '@angular/router';
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
 import { DirectiveWayComponent } from './directive-way/directive-way.component';
 import { FileGeneratorComponent } from './file-generator/file-generator.component';
 import { MultiServiceComponent } from './multi-service/multi-service.component';
@@ -6,7 +7,6 @@ import { MultipleDirectiveComponent } from './multiple-directive/multiple-direct
 import { OnPushComponent } from './on-push/on-push.component';
 import { ServiceCodeWayComponent } from './service-code-way/service-code-way.component';
 import { ServiceWayComponent } from './service-way/service-way.component';
-import { StandaloneComponent } from './standalone/standalone.component';
 import { TusComponent } from './tus/tus.component';
 
 export const appRoutes: Routes = [
@@ -14,10 +14,15 @@ export const appRoutes: Routes = [
   { path: 'directive-way', component: DirectiveWayComponent },
   { path: 'service-way', component: ServiceWayComponent },
   { path: 'service-code-way', component: ServiceCodeWayComponent },
-  { path: 'standalone', component: StandaloneComponent },
   { path: 'on-push', component: OnPushComponent },
   { path: 'multi', component: MultipleDirectiveComponent },
   { path: 'multi2', component: MultiServiceComponent },
   { path: 'tus', component: TusComponent },
   { path: 'file-generator', component: FileGeneratorComponent }
 ];
+
+@NgModule({
+  imports: [RouterModule.forRoot(appRoutes)],
+  exports: [RouterModule]
+})
+export class AppRoutingModule {}
