+++ src/app/app.html
@@ -30,7 +30,6 @@
   <div class="links">
     <span>Select demo:</span>
     <a routerLink="/directive-way" routerLinkActive="active">directive</a>
-    <a routerLink="/standalone" routerLinkActive="active">standalone</a>
     <a routerLink="/service-way" routerLinkActive="active">mixed</a>
     <a routerLink="/service-code-way" routerLinkActive="active">service</a>
     <a routerLink="/on-push" routerLinkActive="active">on-push</a>
